+++ src/term.zig
@@ -1,10 +1,10 @@
 const Variable = @import("variable.zig").Variable;
 
 pub const Term = struct {
-    coefficient: f32,
+    coefficient: f64,
     variable: *Variable,
 
-    pub fn init(coefficient: f32, variable: *Variable) Term {
+    pub fn init(coefficient: f64, variable: *Variable) Term {
         return .{ .coefficient = coefficient, .variable = variable };
     }
 };
